+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Add demo data archives as TOML files to repository
 * Fix resizing PDF viewer loses mouse focus in Firefox
 * Add raptor mascot images as to empty pages
+* Increase contrast of nested form fields
 
 
 ## v2024.3 - 2024-01-09